chore: init refacto options
